Refactor code login with google
