+++ frontend/src/components/LoginWithSocial/LoginWithSocial.jsx
@@ -1,0 +1,33 @@
+import React from 'react';
+import { useState, useEffect } from 'react';
+import StyledFirebaseAuth from 'react-firebaseui/StyledFirebaseAuth';
+import { uiConfig } from '@/configs/firebaseConfig';
+import firebase from 'firebase/compat/app';
+
+const LoginWithSocial = () => {
+  const [isSignedIn, setIsSignedIn] = useState(false);
+
+  useEffect(() => {
+    const unregisterAuthObserver = firebase.auth().onAuthStateChanged((user) => {
+      setIsSignedIn(!!user);
+    });
+    return () => unregisterAuthObserver();
+  }, []);
+
+  if (!isSignedIn) {
+    return (
+      <div>
+        <StyledFirebaseAuth uiConfig={uiConfig} firebaseAuth={firebase.auth()} />
+      </div>
+    );
+  }
+  return (
+    <div>
+      <h1>My App</h1>
+      <p>Welcome {firebase.auth().currentUser.displayName}! You are now signed-in!</p>
+      <a onClick={() => firebase.auth().signOut()}>Sign-out</a>
+    </div>
+  );
+};
+
+export default LoginWithSocial;
