+++ frontend/src/components/Modal/Modal.jsx
@@ -7,7 +7,7 @@ const Modal = ({ isShowing, hide, children, header, footer }) =>
         <React.Fragment>
           <div onClick={hide} className="fixed top-0 left-0 right-0 bottom-0 bg-black opacity-30" />
           <div
-            className="bg-white text-center relative z-50 max-w-[400px] mx-auto rounded-sm top-1/2 -translate-y-3/4"
+            className="bg-white text-center absolute left-1/2 -translate-x-1/2 z-50 min-w-[400px] mx-auto rounded-sm top-1/2 -translate-y-3/4"
             aria-modal
             aria-hidden
             tabIndex={-1}
