+++ frontend/src/configs/firebaseConfig.jsx
@@ -1,0 +1,18 @@
+import firebase from 'firebase/compat/app';
+import 'firebase/compat/auth';
+
+const config = {
+  apiKey: import.meta.env.VITE_REACT_APP_FIREBASE_API_KEY,
+  authDomain: import.meta.env.VITE_REACT_APP_FIREBASE_AUTH_DOMAIN
+};
+firebase.initializeApp(config);
+
+const uiConfig = {
+  signInFlow: 'popup',
+  signInOptions: [firebase.auth.GoogleAuthProvider.PROVIDER_ID],
+  callbacks: {
+    signInSuccessWithAuthResult: () => false
+  }
+};
+
+export { uiConfig };
