+++ frontend/src/pages/Login/Login.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Helmet } from 'react-helmet';
 import { useTranslation } from 'react-i18next';
-import SignInScreen from '../../configs/firebase';
-import LoginFormModal from '../../containers/Modal/LoginFormModal';
+import LoginWithSocial from '@/components/LoginWithSocial/LoginWithSocial';
+import LoginFormModal from '@/containers/Modal/LoginFormModal';
 
 const Login = () => {
   const { t } = useTranslation();
@@ -12,7 +12,7 @@ const Login = () => {
       <Helmet>
         <title>{t('loginText')}</title>
       </Helmet>
-      <div className="container h-[720px] bg-[#f0f2f5] flex items-center justify-center gap-20">
+      <div className="container h-[720px] bg-[#f0f2f5] flex items-center justify-center gap-20 flex-col">
         <div>
           <img className="h-[70px] -mx-5" src="https://static.xx.fbcdn.net/rsrc.php/y8/r/dF5SId3UHWd.svg" alt="logo" />
 
@@ -21,8 +21,9 @@ const Login = () => {
 
           <LoginFormModal />
         </div>
+        <div>{t('or')}</div>
         <div>
-          <SignInScreen />
+          <LoginWithSocial />
         </div>
       </div>
     </>
